removed double inclusion of PCL in cmakelists, something on the lines dunno rememba what

+++ g2o_frontend/line_extraction/viewerGUI.h
@@ -44,8 +44,8 @@ class ViewerGUI : public QMainWindow, public Ui::MainWindow
 
 	public slots:
 	 void updateVal1(int val);
-	 void updateVal2(float val);
-	 void updateVal3(float val);
+	 void updateVal2(int val);
+	 void updateVal3(int val);
 	 void showOriginal();
 	 void lineExtraction();
 	 void setAlgorithm();
